// src/sync/korea/korea-stock.interface.ts
// Generated by https://quicktype.io

export interface KoreaStockXML {
  stockprice: Stockprice;
}

export interface Stockprice {
  $: StockpriceClass;
  TBL_DailyStock: TBLDailyStock[];
  TBL_AskPrice: TBLAskPrice[];
  TBL_StockInfo: TBLStockInfo[];
  TBL_Hoga: TBLHoga[];
  TBL_TimeConclude: StockpriceTBLTimeConclude[];
  stockInfo: StockInfo[];
}

export interface StockpriceClass {
  querytime: string;
}

export interface TBLAskPrice {
  AskPrice: AskPriceElement[];
}

export interface AskPriceElement {
  $: AskPrice;
}

export interface AskPrice {
  member_memdoMem: string;
  member_memdoVol: string;
  member_memsoMem: string;
  member_mesuoVol: string;
}

export interface TBLDailyStock {
  DailyStock: DailyStock[];
}

export interface DailyStock {
  $: { [key: string]: string };
}

export interface TBLHoga {
  $: { [key: string]: string };
}

export interface TBLStockInfo {
  $: MyStockInfo;
}

export interface MyStockInfo {
  JongName: string; //종목명
  CurJuka: string; //현재가
  DungRak: string; //부호처리
  Debi: string; //전일대비
  PrevJuka: string; //전일종가
  Volume: string; //거래량
  Money: string; //거래대금
  StartJuka: string; //시가
  HighJuka: string; //고가
  LowJuka: string; //저가
  High52: string; //52주 최고
  Low52: string; //52주 최저
  UpJuka: string; //상한가
  DownJuka: string; //하한가
  Per: string;
  Amount: string; //상장주식수
  FaceJuka: string; //액면가
}

export interface StockpriceTBLTimeConclude {
  TBL_TimeConclude: TBLTimeConcludeTBLTimeConclude[];
}

export interface TBLTimeConcludeTBLTimeConclude {
  $: TBLTimeConclude;
}

export interface TBLTimeConclude {
  time: string;
  negoprice: string;
  Dungrak: string;
  Debi: string;
  sellprice: string;
  buyprice: string;
  amount: string;
}

export interface StockInfo {
  $: { [key: string]: string };
}

export interface KoreaStockInfoFromCsv {
  code: string;
  krName: string;
  enName: string;
  ipoDate: string;
  exchange: string;
}